requirements.txt
# 基础依赖
numpy>=1.20.0
matplotlib>=3.4.0

# GUI和图形处理
tk>=0.1.0
python-tk>=3.8.0
Pillow>=9.0.0  # 用于图像处理和图标生成

# 打包工具
pyinstaller>=5.0.0
setuptools>=65.5.1
wheel>=0.40.0

# 开发工具
pip>=22.0.0
